Move assignment creating function to new module
